Refacor core logic structure
the code structure has been refactored to fit the platformio
build system and allow for a modular system.
the functionality of each component has remained largely the same
what has changed is how the code interacts with itself, what gets
exposed and the syntax for the code, which has been modified to
better allign with recognizable standards such as the embeded
systems best practices and the linux kernel documentation.

the logic for heach hardware component has been moved to its
own module, each with the corresponding .cpp and .h in the respective
src/ and include/ directories
the idea is to have each module of a harware componente contain
its own private implementation, handling the pins defined in config.h,
implementing system functions and helper methods that are hidding
from the outside. The only internal methods to be exposed are the
ones that the .h "interface" allows.

the display module is a special case of this; it allows for support
of two display types, either an oled or lcd display.
to archieve this each type_display.cpp implements its own libraries
keeping them separate from any of the other modules that don't
require it, they also handle each method and logic inside, exposing
only whats needed from the rest system via their respective .h
the display.cpp then acts as a proper interface, routing the logic
to the implementation that has been selected by the user via
the USE_OLED build flag. (a env system for platformio could be
preffered for a nicer user experience, but its fine for now).

core and kernel have been merged in a single file for the following
reasons:
1. kernel only contained the helper methods that core needed,
which should stay hidden inside, having a separate file means
exposing the methods via .h so that the core can access them which
defeates the whole purpose
2. core and kernel where confusing denominations of what practically
is a finite state machine (hence fsm.cpp as unified name) that
handles the state of system, delegating the single operations
to the corresponding module.
3. for the above reason, the switch condition, that defines
the fsm opeartions and is in turn its core, has been moved to
the fsm.cpp file, exposing just the update method which concernes
the main.cpp loop. this has the added benefit of subtracting
responsability from the arduino component, managing everything
in its own separate space and having the entry point only be concerned
with high level setups and updates. (the io file avoid having to call
each component inside the main.cpp, its not strictly needed but makes
things cleaner overall).

the syntax has been updated to follow the guidelines for the c
language that the linux kernel documentation suggest, this in turn
works, as one of the objectives of this project is to limit the
use of opp, meaning a C wrapper to the components internal
implementations is needed, and in turn makes the syntax more readable
and standardized.