Fix button 1 ghost click
the resulted as pressed after about 115 iteration of the loop
area of the code, causing the button to not be usable for any logic.
the bug origin is the B1 definition used for the button pin, which
happens to be reserved by the Arduino binary.h header that sets said
variable to 1 resulting in a ghost click.
to prevent this a more distinct name for the button pins has been used.
(i've spent a whole day on this bug...)