Fix display saturating ram
switched to the u8g2 library for lighter sram use via the u8x8 module
which doesn't buffer the display data on the device ram, alowing just
for text and bitmap but it being all the device need is well suited
and in turn avoids the issues with the adafruit library which saturated
the devices ram (arduino uno only has 2kb) by using about 1kb
just for the display buffer and making the device glitch out when
more than 3 dynamic sprints where present. in the effort of saving
ram, before moving to the new library, i tried to move the string
data that didn't have to change, the messages that pop up during
the game that aren't user dependant, to the arduino's flash memory.
this helped and allowed for the program to run properly but it wasn't
enough. some other improvements could be done to the overall code
but will tacle that at a later time. sram testing shows that the usage
has dropped significantly, dipping below 150b makes the arudino uno
crash, the string optimizations allowed for about 222b of free space
while the library change showed about 1229b of free space, meaning
the whole program fits in about 800bytes.
as stated further improvements could be had in other critical parts
of the software if other issues arise.