Add round logic
rough commit for tracking purposes, wouldn't commit comments if
i had tracked them before (my bad) will clean later.

the logic for the round is quite simple, we have to match the user
inputed sequence as he goes against the displayed sequence.
we leverage the generated sequence to validate the user input against
it as the user presses the buttons, thus throwing the lose state
transition as soon as it happens.

the logic sits in a "game" module, a better name should be given
later on, but it helps to have the game specific logic separated
from the state management stuff that the fsm hold, this way it has
its proper place of standing.
some furhter logic refactoring should be had so the game module doesn't
have to require other modules, say the button one, which would be
better for code separation and responsibility stuff.

the sequence is just the 4 numbers and gets shuffled each time
a round starts, it uses what it reads from pin 0 which is quite neat.

the fsm logic is still in the testing phase, routing isn't final,
also the ram testing method is still there, will remove later.