Fix buffer size for static print
was using the wrong method for calculating the lenght of the string
that was stored in the arduino flash via progmem. using the proper
method fixes the buffer size issue, also it scales to the string
it receives so it makes more sense for sizing.
finally simplified the method a bit as it was unecessarly convoluted
and renamed from "static" to "P" as this way it mathces the naming
convention from the pgmspace.h that provides the methods for string
interaction with the flash memory (makes more sense than using static)