Add final game loop
fineshed the game loop routing in the fsm, all states switch acording
to the provided spec, the game is playable as is.
some stuff has to be moved to the proper method, the sleep and
potentiometer still have to be implemented properly but the core
game loop logic is sound and the game is playable at a "1 difficulty".

some refactoring will be done to have a more readable and separated code