Add arg support to print_P for score
Static print or progmem print can now accept args for filling
templates, this way static strings can also be composed if need be,
in this case its for the score but a round string could be setup
to ditch the different print methods.
Using args was the most straight forward way of keeping the original
progmem system while adding support for "composed" strings, where one
part is static and the another dynamic. Other solutions could have
been either a getter, which would have required a static buffer
thus filling memory, and another would have been a yet another print
method. This may not be the most optimal or efficient solution but
it was the most practical as of now