Update potentiometer logic
pot defines difficulty factor by scaling the time for each round via
a factor that then defines the diffuclty. the range for the factor
is between 1 and 4, with 1 being the lowest and 4 the highest.
the diffuctulty logic is handled the the fsm for now, could be
moved to the game if need be