Refactor fsm logic + cleanup
a bit more readable, functions stay the same, just a nicer
flow logic for reading the code and having states that are more
self contained.

each state now starts in a "clean state" meaning all leds are off
and all the buttons are not pressed, this avoids having to do the
same resets at entry, those are handled when transitioning so
that a sensible default is set.

cleaned up some comments and logic